+++ src/Ehann/Redisearch/Document/Document.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Eeh\RediSearch\Document;
+namespace Ehann\RediSearch\Document;
 
-use Eeh\RediSearch\Exceptions\FieldNotInSchemaException;
-use Eeh\RediSearch\Fields\FieldFactory;
-use Eeh\RediSearch\Fields\FieldInterface;
+use Ehann\RediSearch\Exceptions\FieldNotInSchemaException;
+use Ehann\RediSearch\Fields\FieldFactory;
+use Ehann\RediSearch\Fields\FieldInterface;
 
 class Document
 {