+++ tests/Stubs/TestDocument.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Eeh\Tests\Stubs;
+namespace Ehann\Tests\Stubs;
 
-use Eeh\RediSearch\Document\Document;
-use Eeh\RediSearch\Fields\NumericField;
-use Eeh\RediSearch\Fields\TextField;
+use Ehann\RediSearch\Document\Document;
+use Ehann\RediSearch\Fields\NumericField;
+use Ehann\RediSearch\Fields\TextField;
 
 /**
  * @property TextField title