+++ docs-src/changelog.md
@@ -3,5 +3,6 @@
 
 ## 0.5.0
 
+* Rename vendor namespace from **Eeh** to **Ehann**
 * **AbstractIndex** was renamed to **Index** and is no longer abstract.
 * Custom document ID is now properly set when adding to an index.